OLM client: Wait till the deployment appear
Sometimes OLM installation fails with below error:
```
INFO[0028] Waiting for deployment/packageserver rollout to complete
FATA[0028] Failed to install OLM version "latest": deployment/packageserver failed to rollout: deployments.apps "packageserver" not found# Please enter the commit message for your changes. Lines starting
```

The reason might be that the deployment object we are waiting for is not
yet registered with the API server. So, do not treat this as an error
and continue the polling.